refactor(Cell): 셀 frame 파일 분리
